Create README.md
Cautiously, we can say that Dynamic Island is the invention of a new way of interacting with the user on a par with push notifications on the iPhone and TouchBar on the Mac. Therefore, it is extremely interesting how it will work: the functioning of the island directly depends on how application developers will implement some functions into it. And, as we know from experience, the TouchBar did not fly, although it also made noise at the time. There is no information for developers about Dynamic Island on the Apple website yet, so we will wait for more data on what information can be placed on the island for application developers and what actions users will be able to perform.

Have you already imagined how you could use the island for your application? Write in the comments!